Refactored structure of DefinitionsPageModule
