--- projects/app/src/app/pages/definitions/components/index.ts
@@ -1,1 +1,0 @@
-export * from './definitions/definitions.component';
