+++ projects/app/src/app/pages/definitions/pages/definitions/definitions-routing.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+import { DefinitionsPageComponent } from './definitions.component';
+
 const routes: Routes = [
   {
     path: '',
-    loadChildren: () => import('./pages/dashboard/dashboard.module').then(m => m.DashboardModule)
+    component: DefinitionsPageComponent
   }
 ];
 
@@ -12,4 +14,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class DashboardPageRoutingModule {}
+export class DefinitionsRoutingModule {}
