+++ projects/app/src/app/pages/definitions/pages/definitions/definitions.component.scss
