+++ projects/app/src/app/pages/definitions/pages/definitions/index.ts
@@ -1,0 +1,4 @@
+export * from './definitions.component';
+
+export * from './definitions-routing.module';
+export * from './definitions.module';
