+++ projects/app/src/app/pages/definitions/pages/index.ts
@@ -1,0 +1,1 @@
+export * from './definitions';
